refactor: Crate a separeted package for User DTOs into dtos package
